+++ compiler/CHANGELOG.md
@@ -4,6 +4,10 @@ This changelog summarizes newly introduced optimizations and other compiler rela
 
 ## GraalVM for JDK 26 (Internal Version 26.0.0)
 * (GR-69280): Allow use of the `graal.` prefix for Graal compiler options without issuing a warning.
+* (GR-58163): Added support for recording and replaying JIT compilations. The `-Djdk.graal.RecordForReplay=*` option
+  serializes all compilations matching the pattern to JSON files, which contain the results of JVMCI calls. The
+  recorded compilations can be replayed with the `mx replaycomp` command. Truffle compilations are currently not
+  supported. See `docs/ReplayCompilation.md` for details.
 
 ## GraalVM 25 (Internal Version 25.0.0)
 * (GR-60088): This PR adds the `org.graalvm.nativeimage.libgraal` SDK module. With this module, all logic for building